# internal/transactions/adapters/postgres/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type SchemaMigration struct {
	Version string `json:"version"`
}

type Transaction struct {
	ID              pgtype.UUID        `json:"id"`
	UserID          string             `json:"user_id"`
	Amount          int64              `json:"amount"`
	TransactionType string             `json:"transaction_type"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
}

type TransactionsDefault struct {
	ID              pgtype.UUID        `json:"id"`
	UserID          string             `json:"user_id"`
	Amount          int64              `json:"amount"`
	TransactionType string             `json:"transaction_type"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
}

type TransactionsY2026M04 struct {
	ID              pgtype.UUID        `json:"id"`
	UserID          string             `json:"user_id"`
	Amount          int64              `json:"amount"`
	TransactionType string             `json:"transaction_type"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
}

type TransactionsY2026M05 struct {
	ID              pgtype.UUID        `json:"id"`
	UserID          string             `json:"user_id"`
	Amount          int64              `json:"amount"`
	TransactionType string             `json:"transaction_type"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
}

type TransactionsY2026M06 struct {
	ID              pgtype.UUID        `json:"id"`
	UserID          string             `json:"user_id"`
	Amount          int64              `json:"amount"`
	TransactionType string             `json:"transaction_type"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
}

type TransactionsY2026M07 struct {
	ID              pgtype.UUID        `json:"id"`
	UserID          string             `json:"user_id"`
	Amount          int64              `json:"amount"`
	TransactionType string             `json:"transaction_type"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
}